No more serialized paths
I renamed mod.path to mod.identifier.

Now, it ought to be some kind of unique identifier -- we could go with what's
in mods' mod_info.lua, but it's neat to have kind canonical short names for
often used mods (faf, nomads), so I'm not sure what to go with.