wine Mumble hack - initial commit
Adds code that detects whether FAF was launched uner wine. Uses it to
hack together a method to pass mumble link calls to Linux Mumble.

The code launches a Python script linux-side and loads a mumble_link
stub module that passes calls to it through sockets. It needs an
appropriate mumble_link linux library to run.

I didn't have the means to test the code, but it's syntactically correct
and shouldn't break the windows build.