Use a GITHUB_TOKEN for Github API in appveyor
This changes the github api access (for finding latest faf-uid version)
to using authenticated requests.
It uses env vars called `GITHUB_USER` and `GITHUB_TOKEN` that must be filled with a
Github Personal Access Token. Those can be added in Appveyor project
settings in "Environment" or specified in the build.

This commit adds a token for fafbot to the build.