add subdirectory support to Local Archive
Sub directories in the local replays folder will be listed as 'directory'
items. Doubleclick on one of the subfolders will show the content of that
subfolder as current replay folder.
If the current folder is a sub folder (not util.REPLAY_DIR) a directory up
element '..' is added. A doubleclick will show the
parent folder again as current replay folder.

add closed_folder & open_folder icons