Add user class
Add a class that abstracts a user. This will keep the player that
the user is, keep his friends / foes / etc. It also never gets
replaced or destroyed, meaning that in the future we'll be able to just
pass it in the constructor wherever and never worry about it again.

Signed-off-by: Igor Kotrasinski <ikotrasinsk@gmail.com>