Decouple update connection from updater
Have the update connection use methods of a mixin class to notify the
updater about read events and state changes - since there should be only
one user of the read data, we don't need to use signals for that.
This decouples talking to the server from the updater itself and lays
some groundwork towards unifying our simple protocol in a single class.

Signed-off-by: Igor Kotrasinski <ikotrasinsk@gmail.com>