Refactor automatic login
Automatic login is now governed by a separate variable that we can
enable (if we successfully logged in, for example) or disable (if we
failed to login or received a critical server error).

We also move logging in by either login window or automatically to a
single method, and call it instead of directly showing the window or
doing autologin.

Signed-off-by: Igor Kotrasinski <ikotrasinsk@gmail.com>