Use transactions to control model update signals
All model update operations now accept a _transaction argument that can
be used to delay emitting signals until the model update is complete and
the state is consistent again. New signals are added for the purpose of
notifying other parts of a model that something changed (instead of
using normal update signals like we did before).

This way we never have to worry about signals being emitted too early
again.

Signed-off-by: Igor Kotrasinski <ikotrasinsk@gmail.com>